backend/project-director/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
coverage==5.4
cryptography==3.3.1
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.4
django-appconf==1.0.4
django-googledrive-storage==1.5.2
django-redis==4.12.1
djangorestframework==3.12.2
django-cors-headers
djangorestframework-simplejwt
filelock==3.0.12
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
#mod_wsgi==4.9.0
Pillow==8.0.1
protobuf==3.14.0
#psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.2.2
wrapt==1.12.1
